Enforce slightly different behavior when running as root/daemon.
For root users:
   o Daemonize by default, use -n to prevent this
   o Enable syslog instead of stdout/stderr
   o Use adjtime(), -l flag, by default
   o Disable RFC4330 packet validation, -t, by default

For regular users there is no change, rather the opposite.  Changes
made prior to this commit are reverted for "user mode" to make sure
the same behavior as previous releases is maintained.

Debug mode is now always available, the ENABLE_DEBUG compile flag,
does however enable even more debug options.  But to see any of them
you still need to supply the -d flag.

Added a -V flag to display version and copyright information, as well
as extended the usage listing to include a full help listing with a
option legend.

Update documentation accordingly and for new features.